+++ CHANGELOG.md
@@ -29,6 +29,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 be run with the `fpm run --example` command.
 - `build.sh` improved.
 - In examples, keyword argument lists are now generally used to improve understanding.
+- The multi-tracks examples `canon.f90`, `blues.f90` and `la_folia.f90` now uses the control change Pan to obtain a stereo effect.
 - The API was simplified by renaming many methods and arguments.
 - The method `Note()` was split in two: `Note_ON()` and `Note_OFF()`.